requirements.txt
Appium-Python-Client==4.0.1
attrs==23.2.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
exceptiongroup==1.2.1
h11==0.14.0
idna==3.7
iniconfig==2.0.0
jinja2==3.1.4
MarkupSafe==2.1.5
nodeenv==1.9.1
outcome==1.3.0.post0
packaging==24.1
pip==21.1.3
pluggy==1.5.0
pycparser==2.22
PySocks==1.7.1
pytest==8.2.2
pytest-base-url==2.1.0
pytest-html==4.1.1
pytest-metadata==3.1.1
pytest-selenium==4.1.0
pytest-variables==3.1.0
requests==2.32.3
selenium==4.22.0
setuptools==56.0.0
sniffio==1.3.1
sortedcontainers==2.4.0
tenacity==8.5.0
tomli==2.0.1
trio==0.26.0
trio-websocket==0.11.1
typing-extensions==4.12.2
urllib3==2.2.2
websocket-client==1.8.0
wsproto==1.2.0
